requirements.txt
python-dotenv
faiss-cpu
bittensor
pydantic
fastapi
numpy
pytorch
weaviate-client
scikit-learn
sentence-transformers